Little pattern machine integration (#118)
This is a major revision of Scamper to support a newly designed back-end, the Little Pattern Machine (LPM). LPM is a core calculus for core functional programs featuring pattern matching and realized as a tiny VM. The front-end of Scamper has also been refactored to accommodate different languages in the future. Finally, the UI was also revamped to be a single interface/IDE with a built-in file chooser rather than having file choosing be a separate page.